fix: remove warning about deprecated set-output
The old version of actions core uses set-output command which is deprecated.  In my build I am getting the following warnings:

```
Warning: The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/
```

This bumps the version of actions/core to a new version that uses Environment files and thus gets rid of this warning.